+++ test/torch/model/test_deepar_modules.py
@@ -17,45 +17,6 @@ import pytest
 import torch
 
 from gluonts.torch.model.deepar import DeepARLightningModule, DeepARModel
-from gluonts.torch.model.deepar.util import lagged_sequence_values
-
-
-@pytest.mark.parametrize(
-    "lag_indices, prior_sequence, sequence",
-    [
-        (
-            [0, 1, 5, 10, 20],
-            torch.randn((4, 100)),
-            torch.randn((4, 8)),
-        ),
-        (
-            [0, 1, 5, 10, 20],
-            torch.randn((4, 100, 1)),
-            torch.randn((4, 8, 1)),
-        ),
-        (
-            [0, 1, 5, 10, 20],
-            torch.randn((4, 100, 2)),
-            torch.randn((4, 8, 2)),
-        ),
-    ],
-)
-def test_lagged_sequence_values(
-    lag_indices: List[int],
-    prior_sequence: torch.Tensor,
-    sequence: torch.Tensor,
-):
-    res = lagged_sequence_values(lag_indices, prior_sequence, sequence, None)
-    full_sequence = torch.cat((prior_sequence, sequence), dim=1)
-    for t in range(res.shape[1]):
-        expected_lags_t = torch.stack(
-            [
-                full_sequence[:, t + prior_sequence.shape[1] - l]
-                for l in lag_indices
-            ],
-            dim=-1,
-        ).reshape(sequence.shape[0], -1)
-        assert torch.allclose(expected_lags_t, res[:, t, :])
 
 
 @pytest.mark.parametrize(
